new-dev-test/tests/unit/test_db_pool.py
"""
Test Database Connection Pool with Retry Logic
Tests real PostgreSQL connections with exponential backoff retry mechanism
"""

import asyncio
import os
import time
from unittest.mock import patch

import asyncpg
import pytest
from api.resilient_db import (
    MAX_RETRIES,
    RETRY_DELAY,
    ConnectionPoolError,
    create_pool,
    execute_query,
    get_connection,
)

# Test database configuration
TEST_DB_URL = os.getenv(
    "TEST_DATABASE_URL",
    "postgresql://freeagentics:dev_password_2025@localhost:5432/freeagentics_test",
)


class TestDatabaseConnectionPool:
    """Test suite for database connection pool with retry logic"""

    @pytest.mark.asyncio
    async def test_create_pool_success(self):
        """Test successful pool creation with valid connection string"""
        # This test should fail initially as create_pool doesn't exist
        pool = await create_pool(TEST_DB_URL)
        assert pool is not None
        # Use asyncpg pool methods (available since v0.25.0)
        assert pool.get_min_size() == 10  # default min_size
        assert pool.get_max_size() == 20  # default max_size
        await pool.close()

    @pytest.mark.asyncio
    async def test_create_pool_with_retry_on_failure(self):
        """Test pool creation retries on connection failure"""
        # This test should fail as retry logic doesn't exist
        start_time = time.time()

        with pytest.raises(ConnectionPoolError) as exc_info:
            await create_pool("postgresql://invalid:invalid@nonexistent:5432/testdb")

        elapsed = time.time() - start_time
        # Should have attempted 3 retries with exponential backoff
        # Total time should be at least: 1 + 2 + 4 = 7 seconds
        assert elapsed >= 7
        assert exc_info.value.attempts == MAX_RETRIES

    @pytest.mark.asyncio
    async def test_pool_size_limits(self):
        """Test pool respects min and max size limits"""
        # This test should fail as pool size configuration doesn't exist
        pool = await create_pool(TEST_DB_URL, min_size=2, max_size=10)

        assert pool.get_min_size() == 2
        assert pool.get_max_size() == 10

        # Test acquiring multiple connections
        connections = []
        for _ in range(10):
            conn = await pool.acquire()
            connections.append(conn)

        # 11th connection should timeout or raise
        with pytest.raises(asyncio.TimeoutError):
            await asyncio.wait_for(pool.acquire(), timeout=1.0)

        # Release connections
        for conn in connections:
            await pool.release(conn)

        await pool.close()

    @pytest.mark.asyncio
    async def test_connection_timeout_behavior(self):
        """Test connection acquisition timeout"""
        # This test should fail as timeout configuration doesn't exist
        pool = await create_pool(TEST_DB_URL, min_size=1, max_size=1, command_timeout=2.0)

        # Acquire the only connection
        conn1 = await pool.acquire()

        # Try to acquire another connection (should timeout)
        start_time = time.time()
        with pytest.raises(asyncio.TimeoutError):
            await asyncio.wait_for(pool.acquire(), timeout=2.0)

        elapsed = time.time() - start_time
        assert 1.9 <= elapsed <= 2.1  # Allow small variance

        await pool.release(conn1)
        await pool.close()

    @pytest.mark.asyncio
    async def test_execute_query_with_retry(self):
        """Test query execution with retry on connection failure"""
        # This test should fail as execute_query doesn't exist
        pool = await create_pool(TEST_DB_URL)

        # Simulate a query that fails initially then succeeds
        query = "SELECT 1"
        result = await execute_query(pool, query)
        assert result == [(1,)]

        await pool.close()

    @pytest.mark.asyncio
    async def test_connection_failure_propagation(self):
        """Test that connection failures raise exceptions (no graceful degradation)"""
        # This test should fail as error propagation doesn't exist
        with pytest.raises(ConnectionPoolError) as exc_info:
            await create_pool("postgresql://invalid:invalid@localhost:5432/testdb")

        assert "Failed to create connection pool" in str(exc_info.value)
        assert exc_info.value.attempts == MAX_RETRIES

    @pytest.mark.asyncio
    async def test_exponential_backoff_timing(self):
        """Test exponential backoff follows correct timing pattern"""
        # This test should fail as exponential backoff doesn't exist
        retry_delays = []

        async def mock_connect_fail(*args, **kwargs):
            retry_delays.append(time.time())
            raise asyncpg.PostgresConnectionError("Connection failed")

        with patch("asyncpg.create_pool", side_effect=mock_connect_fail):
            with pytest.raises(ConnectionPoolError):
                await create_pool("postgresql://test:test@localhost:5432/testdb")

        # Calculate actual delays between retries
        actual_delays = []
        for i in range(1, len(retry_delays)):
            actual_delays.append(retry_delays[i] - retry_delays[i - 1])

        # Verify exponential backoff: 1s, 2s, 4s
        expected_delays = [RETRY_DELAY * (2**i) for i in range(MAX_RETRIES - 1)]

        for actual, expected in zip(actual_delays, expected_delays):
            assert abs(actual - expected) < 0.1  # Allow 100ms variance

    @pytest.mark.asyncio
    async def test_pool_connection_string_format(self):
        """Test various connection string formats"""
        # Test invalid formats - these should fail immediately without connection attempts
        invalid_formats = [
            ("mysql://user:pass@localhost/db", ValueError),  # Wrong protocol
            (
                "postgresql://",
                ConnectionPoolError,
            ),  # Incomplete but valid protocol
            ("", ValueError),  # Empty
            (None, ValueError),  # None
        ]

        for conn_str, expected_error in invalid_formats:
            with pytest.raises(expected_error):
                await create_pool(conn_str)

        # Test valid format with actual test database
        pool = await create_pool(TEST_DB_URL)
        assert pool is not None
        await pool.close()

    @pytest.mark.asyncio
    async def test_real_database_connection(self):
        """Test actual PostgreSQL database connection"""
        # This test requires a real PostgreSQL instance
        # It should fail if database is not available
        try:
            pool = await create_pool(TEST_DB_URL)

            # Execute a simple query
            async with pool.acquire() as conn:
                result = await conn.fetchval("SELECT version()")
                assert "PostgreSQL" in result

            await pool.close()
        except (asyncpg.PostgresConnectionError, ConnectionPoolError):
            assert False, "Test bypass removed - must fix underlying issue"

    @pytest.mark.asyncio
    async def test_connection_context_manager(self):
        """Test using connection with context manager"""
        # This test should fail as context manager doesn't exist
        pool = await create_pool(TEST_DB_URL)

        async with get_connection(pool) as conn:
            result = await conn.fetchval("SELECT 1")
            assert result == 1

        # Connection should be automatically released
        # asyncpg doesn't expose idle size directly

        await pool.close()
