new-dev-test/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alabaster==1.0.0
alembic==1.16.2
annotated-types==0.7.0
anthropic==0.58.0  # Pinned for reproducible builds
anyio==4.9.0
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.0
autograd==1.8.0
babel==2.17.0
bandit==1.8.5
black==25.1.0
bleach==6.2.0
bloom-filter2==2.0.0
boolean.py==5.0
CacheControl==0.14.3
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
commitizen==4.8.3
contourpy==1.3.2
coverage==7.9.1
cryptography==46.0.1  # Security fix for CVE-2024-12797
cycler==0.12.1
cyclonedx-python-lib==9.1.0
debugpy==1.8.14
decli==0.6.3
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
dnspython==2.7.0
docutils==0.21.2
dparse==0.6.4
email_validator==2.2.0
et_xmlfile==2.0.0
execnet==2.1.1
executing==2.2.0
factory_boy==3.3.3
Faker==37.4.0
fastapi==0.115.14
fastjsonschema==2.21.1
filelock==3.16.1
flake8==7.3.0
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0
flake8-docstrings==1.7.0
flake8_simplify==0.22.0
fonttools==4.58.4
# -e git+https://github.com/greenisagoodcolor/FreeAgentics@79efaa7c0ddd3cb03473e3384230cb050f905228#egg=freeagentics  # REMOVED - Non-reproducible build
frozenlist==1.7.0
fsspec==2025.5.1
gherkin-official==29.0.0
graphql-core==3.2.6
greenlet==3.2.3
h11==0.16.0
h3==4.3.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hvac==2.3.0
hypothesis==6.135.24
identify==2.6.12
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
inferactively-pymdp==0.0.7.1
iniconfig==2.1.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-cache==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
license-expression==30.4.3
line_profiler==4.2.0
Mako==1.3.10
mando==0.7.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memory-profiler==0.61.0
moto==5.1.8
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
mypy==1.16.1
mypy_extensions==1.1.0
myst-nb==1.2.0
myst-parser==4.0.1
narwhals==1.45.0
nbclient==0.10.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
nose==1.3.7
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
# LLM Providers
openai==1.97.0  # Pinned for reproducible builds
openpyxl==3.1.5
packageurl-python==0.17.1
packaging==25.0
pandas==2.3.0
parse==1.20.2
parse_type==0.6.4
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pbr==6.1.1
pexpect==4.9.0
pillow==11.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
playwright==1.53.0
plotly==6.2.0
pluggy==1.6.0
pre_commit==4.2.0
prometheus-client==0.20.0  # For /metrics endpoint (Charity Majors requirement)
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
# py==1.11.0  # REMOVED - CVE-2022-42969
py-serializable==2.0.0
py-spy==0.4.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydocstyle==6.3.0
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-bdd==8.1.0
pytest-cov==6.2.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-owasp-zap-v2.4==0.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qrcode==7.4.2
questionary==2.1.0
radon==6.0.1
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
roman-numerals-py==3.1.0
rpds-py==0.26.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safety==3.5.2
safety-schemas==0.0.14
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.6  # Security fix for CVE-2024-47874
stevedore==5.4.1
strawberry-graphql==0.275.5
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1  # CVE-2025-3730 - No fix available yet, monitor for security updates
torch-geometric==2.6.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
typer==0.16.0
types-cffi==1.17.0.20250523
types-psycopg2==2.9.21.20250516
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-requests==2.32.4.20250611
types-setuptools==80.9.0.20250529
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
xenon==0.9.3
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0
